+++ src/ModuleResolver.ts
@@ -23,13 +23,39 @@ import {
   PrettierResolveConfigOptions,
   PrettierVSCodeConfig,
 } from "./types";
-import { getConfig, getWorkspaceRelativePath } from "./util";
+import { getConfig, getWorkspaceRelativePath, isAboveV3 } from "./util";
+import { PrettierWorkerInstance } from "./PrettierWorkerInstance";
+import { PrettierInstance } from "./PrettierInstance";
+import { PrettierMainThreadInstance } from "./PrettierMainThreadInstance";
+import { loadNodeModule, resolveConfigPlugins } from "./ModuleLoader";
 
 const minPrettierVersion = "1.13.0";
-declare const __webpack_require__: typeof require;
-declare const __non_webpack_require__: typeof require;
 
 export type PrettierNodeModule = typeof prettier;
+
+const origFsStatSync = fs.statSync;
+const fsStatSyncWorkaround = (
+  path: fs.PathLike,
+  options: fs.StatSyncOptions
+) => {
+  if (
+    options?.throwIfNoEntry === true ||
+    options?.throwIfNoEntry === undefined
+  ) {
+    return origFsStatSync(path, options);
+  }
+  options.throwIfNoEntry = true;
+  try {
+    return origFsStatSync(path, options);
+  } catch (error: any) {
+    if (error.code === "ENOENT") {
+      return undefined;
+    }
+    throw error;
+  }
+};
+// @ts-expect-error Workaround for https://github.com/prettier/prettier-vscode/issues/3020
+fs.statSync = fsStatSyncWorkaround;
 
 const globalPaths: {
   [key: string]: { cache: string | undefined; get(): string | undefined };
@@ -69,7 +95,8 @@ function globalPathGet(packageManager: PackageManagers): string | undefined {
 export class ModuleResolver implements ModuleResolverInterface {
   private findPkgCache: Map<string, string>;
   private ignorePathCache = new Map<string, string>();
-  private path2Module = new Map<string, PrettierNodeModule>();
+
+  private path2Module = new Map<string, PrettierInstance>();
 
   constructor(private loggingService: LoggingService) {
     this.findPkgCache = new Map();
@@ -79,13 +106,50 @@ export class ModuleResolver implements ModuleResolverInterface {
     return prettier;
   }
 
+  private loadPrettierVersionFromPackageJson(modulePath: string): string {
+    const packageJsonPath = findUp.sync(
+      (dir) => {
+        const pkgFilePath = path.join(dir, "package.json");
+        if (fs.existsSync(pkgFilePath)) {
+          return pkgFilePath;
+        }
+      },
+      { cwd: path.dirname(modulePath) }
+    );
+
+    if (!packageJsonPath) {
+      throw new Error("Cannot find Prettier package.json");
+    }
+
+    const prettierPkgJson = loadNodeModule(packageJsonPath);
+
+    let version: string | null = null;
+
+    if (
+      typeof prettierPkgJson === "object" &&
+      prettierPkgJson !== null &&
+      "version" in prettierPkgJson &&
+      // @ts-expect-error checked
+      typeof prettierPkgJson.version === "string"
+    ) {
+      // @ts-expect-error checked
+      version = prettierPkgJson.version;
+    }
+
+    if (!version) {
+      throw new Error("Cannot load Prettier version from package.json");
+    }
+
+    return version;
+  }
+
   /**
    * Returns an instance of the prettier module.
    * @param fileName The path of the file to use as the starting point. If none provided, the bundled prettier will be used.
    */
   public async getPrettierInstance(
     fileName: string
-  ): Promise<PrettierNodeModule | undefined> {
+  ): Promise<PrettierNodeModule | PrettierInstance | undefined> {
     if (!workspace.isTrusted) {
       this.loggingService.logDebug(UNTRUSTED_WORKSPACE_USING_BUNDLED_PRETTIER);
       return prettier;
@@ -148,7 +212,8 @@ export class ModuleResolver implements ModuleResolverInterface {
       }
     }
 
-    let moduleInstance: PrettierNodeModule | undefined = undefined;
+    let moduleInstance: PrettierInstance | undefined = undefined;
+
     if (modulePath !== undefined) {
       this.loggingService.logDebug(
         `Local prettier module path: '${modulePath}'`
@@ -159,7 +224,16 @@ export class ModuleResolver implements ModuleResolverInterface {
         return moduleInstance;
       } else {
         try {
-          moduleInstance = this.loadNodeModule<PrettierNodeModule>(modulePath);
+          const prettierVersion =
+            this.loadPrettierVersionFromPackageJson(modulePath);
+
+          const isAboveVersion3 = isAboveV3(prettierVersion);
+
+          if (isAboveVersion3) {
+            moduleInstance = new PrettierWorkerInstance(modulePath);
+          } else {
+            moduleInstance = new PrettierMainThreadInstance(modulePath);
+          }
           if (moduleInstance) {
             this.path2Module.set(modulePath, moduleInstance);
           }
@@ -178,20 +252,14 @@ export class ModuleResolver implements ModuleResolverInterface {
     }
 
     if (moduleInstance) {
-      // If the instance is missing `format`, it's probably
-      // not an instance of Prettier
-      const isPrettierInstance = !!moduleInstance.format;
-      const isValidVersion =
-        moduleInstance.version &&
-        !!moduleInstance.getSupportInfo &&
-        !!moduleInstance.getFileInfo &&
-        !!moduleInstance.resolveConfig &&
-        semver.gte(moduleInstance.version, minPrettierVersion);
+      const version = await moduleInstance.import();
 
-      if (!isPrettierInstance && prettierPath) {
+      if (!version && prettierPath) {
         this.loggingService.logError(INVALID_PRETTIER_PATH_MESSAGE);
         return undefined;
       }
+
+      const isValidVersion = version && semver.gte(version, minPrettierVersion);
 
       if (!isValidVersion) {
         this.loggingService.logInfo(
@@ -200,9 +268,7 @@ export class ModuleResolver implements ModuleResolverInterface {
         this.loggingService.logError(OUTDATED_PRETTIER_VERSION_MESSAGE);
         return undefined;
       } else {
-        this.loggingService.logDebug(
-          `Using prettier version ${moduleInstance.version}`
-        );
+        this.loggingService.logDebug(`Using prettier version ${version}`);
       }
       return moduleInstance;
     } else {
@@ -258,8 +324,16 @@ export class ModuleResolver implements ModuleResolverInterface {
     return resolvedIgnorePath;
   }
 
-  public async getResolvedConfig(
-    { fileName, uri }: TextDocument,
+  public async resolveConfig(
+    prettierInstance: {
+      resolveConfigFile(filePath?: string): Promise<string | null>;
+      resolveConfig(
+        fileName: string,
+        options?: prettier.ResolveConfigOptions
+      ): Promise<PrettierOptions | null>;
+    },
+    uri: Uri,
+    fileName: string,
     vscodeConfig: PrettierVSCodeConfig
   ): Promise<"error" | "disabled" | PrettierOptions | null> {
     const isVirtual = uri.scheme !== "file" && uri.scheme !== "vscode-userdata";
@@ -267,14 +341,14 @@ export class ModuleResolver implements ModuleResolverInterface {
     let configPath: string | undefined;
     try {
       if (!isVirtual) {
-        configPath = (await prettier.resolveConfigFile(fileName)) ?? undefined;
+        configPath =
+          (await prettierInstance.resolveConfigFile(fileName)) ?? undefined;
       }
     } catch (error) {
       this.loggingService.logError(
         `Error resolving prettier configuration for ${fileName}`,
         error
       );
-
       return "error";
     }
 
@@ -291,7 +365,7 @@ export class ModuleResolver implements ModuleResolverInterface {
     try {
       resolvedConfig = isVirtual
         ? null
-        : await prettier.resolveConfig(fileName, resolveConfigOptions);
+        : await prettierInstance.resolveConfig(fileName, resolveConfigOptions);
     } catch (error) {
       this.loggingService.logError(
         "Invalid prettier configuration file detected.",
@@ -308,7 +382,7 @@ export class ModuleResolver implements ModuleResolverInterface {
     }
 
     if (resolvedConfig) {
-      resolvedConfig = this.resolveConfigPlugins(resolvedConfig, fileName);
+      resolvedConfig = resolveConfigPlugins(resolvedConfig, fileName);
     }
 
     if (
@@ -322,6 +396,24 @@ export class ModuleResolver implements ModuleResolverInterface {
       );
       return "disabled";
     }
+
+    return resolvedConfig;
+  }
+
+  public async getResolvedConfig(
+    { fileName, uri }: TextDocument,
+    vscodeConfig: PrettierVSCodeConfig
+  ): Promise<"error" | "disabled" | PrettierOptions | null> {
+    const prettierInstance: typeof prettier | PrettierInstance =
+      (await this.getPrettierInstance(fileName)) || prettier;
+
+    const resolvedConfig = await this.resolveConfig(
+      prettierInstance,
+      uri,
+      fileName,
+      vscodeConfig
+    );
+
     return resolvedConfig;
   }
 
@@ -329,71 +421,16 @@ export class ModuleResolver implements ModuleResolverInterface {
    * Clears the module and config cache
    */
   public async dispose() {
-    prettier.clearConfigCache();
+    await prettier.clearConfigCache();
     this.path2Module.forEach((module) => {
       try {
+        // eslint-disable-next-line @typescript-eslint/no-floating-promises
         module.clearConfigCache();
       } catch (error) {
         this.loggingService.logError("Error clearing module cache.", error);
       }
     });
     this.path2Module.clear();
-  }
-
-  private get nodeModuleLoader() {
-    return typeof __webpack_require__ === "function"
-      ? __non_webpack_require__
-      : require;
-  }
-
-  // Source: https://github.com/microsoft/vscode-eslint/blob/master/server/src/eslintServer.ts
-  private loadNodeModule<T>(moduleName: string): T | undefined {
-    try {
-      return this.nodeModuleLoader(moduleName);
-    } catch (error) {
-      this.loggingService.logError(
-        `Error loading node module '${moduleName}'`,
-        error
-      );
-    }
-    return undefined;
-  }
-
-  private resolveNodeModule(moduleName: string, options?: { paths: string[] }) {
-    try {
-      return this.nodeModuleLoader.resolve(moduleName, options);
-    } catch (error) {
-      this.loggingService.logError(
-        `Error resolve node module '${moduleName}'`,
-        error
-      );
-    }
-    return undefined;
-  }
-
-  /**
-   * Resolve plugin package path for symlink structure dirs
-   * See https://github.com/prettier/prettier/issues/8056
-   */
-  private resolveConfigPlugins(
-    config: PrettierOptions,
-    fileName: string
-  ): PrettierOptions {
-    if (config?.plugins?.length) {
-      config.plugins = config.plugins.map((plugin) => {
-        if (
-          typeof plugin === "string" &&
-          !plugin.startsWith(".") &&
-          !path.isAbsolute(plugin)
-        ) {
-          return (
-            this.resolveNodeModule(plugin, { paths: [fileName] }) || plugin
-          );
-        }
-        return plugin;
-      });
-    }
-    return config;
   }
 
   private isInternalTestRoot(dir: string): boolean {