+++ src/StatusBar.ts
@@ -1,4 +1,4 @@
-import { StatusBarAlignment, StatusBarItem, window } from "vscode";
+import { StatusBarAlignment, StatusBarItem, window, ThemeColor } from "vscode";
 
 export enum FormatterStatus {
   Ready = "check-all",
@@ -32,16 +32,24 @@ export class StatusBar {
    */
   public update(result: FormatterStatus): void {
     this.statusBarItem.text = `$(${result.toString()}) Prettier`;
-    // Waiting for VS Code 1.53: https://github.com/microsoft/vscode/pull/116181
-    // if (result === FormattingResult.Error) {
-    //   this.statusBarItem.backgroundColor = new ThemeColor(
-    //     "statusBarItem.errorBackground"
-    //   );
-    // } else {
-    //   this.statusBarItem.backgroundColor = new ThemeColor(
-    //     "statusBarItem.fourgroundBackground"
-    //   );
-    // }
+    switch (result) {
+      case FormatterStatus.Ignore:
+      case FormatterStatus.Warn:
+        this.statusBarItem.backgroundColor = new ThemeColor(
+          "statusBarItem.warningBackground"
+        );
+        break;
+      case FormatterStatus.Error:
+        this.statusBarItem.backgroundColor = new ThemeColor(
+          "statusBarItem.errorBackground"
+        );
+        break;
+      default:
+        this.statusBarItem.backgroundColor = new ThemeColor(
+          "statusBarItem.fourgroundBackground"
+        );
+        break;
+    }
     this.statusBarItem.show();
   }