+++ src/test/suite/plugins.test.ts
@@ -4,7 +4,11 @@ import { format, getText } from "./format.test";
 suite("Test plugins", function () {
   this.timeout(10000);
   test("it formats with plugins", async () => {
-    const { actual } = await format("plugins", "index.php");
+    const { actual } = await format(
+      "plugins",
+      "index.php",
+      /* shouldRetry */ true
+    );
     const expected = await getText("plugins", "index.result.php");
     assert.equal(actual, expected);
   });
@@ -12,6 +16,12 @@ suite("Test plugins", function () {
   test("it correctly resolved plugin in pnpm node_modules dirs structure", async () => {
     const { actual } = await format("plugins-pnpm", "index.js");
     const expected = await getText("plugins-pnpm", "index.result.js");
+    assert.equal(actual, expected);
+  });
+
+  test("it should be able to obtain the `inferredParser` of the plugin", async () => {
+    const { actual } = await format("plugins-pnpm", "index.php");
+    const expected = await getText("plugins-pnpm", "index.result.php");
     assert.equal(actual, expected);
   });
 });