+++ test-fixtures/v3/index.result.ts
@@ -1,0 +1,6 @@
+const Counter = decorator("my-counter")((props: {
+    initialCount?: number;
+    label?: string;
+}) => {
+    // ...
+});